# tests/fixtures/openapi.json
{
  "components": {
    "schemas": {
      "AccountStatus": {
        "enum": [
          "PENDING_APPROVAL",
          "ACTIVE",
          "LOCKED",
          "SUSPENDED",
          "REJECTED",
          "DISABLED"
        ],
        "title": "AccountStatus",
        "type": "string"
      },
...
      "BuyerRegistrationRequest": {
        "additionalProperties": false,
        "description": "Buyer sign-up: identity credentials plus the buyer company profile.\n\nUnexpected fields are rejected (``extra=\"forbid\"``) rather than silently\ndiscarded, so a client cannot smuggle unmodelled data past validation.",
        "properties": {
          "company_name": {
            "maxLength": 255,
            "minLength": 2,
            "title": "Company Name",
            "type": "string"
          },
          "contact_name": {
            "maxLength": 255,
            "minLength": 2,
            "title": "Contact Name",
            "type": "string"
          },
          "country": {
            "maxLength": 100,
            "minLength": 2,
            "title": "Country",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 12,
            "title": "Password",
            "type": "string"
          },
          "phone": {
            "maxLength": 50,
            "minLength": 3,
            "title": "Phone",
            "type": "string"
          },
          "username": {
            "maxLength": 150,
            "minLength": 3,
            "pattern": "^[A-Za-z0-9._-]+$",
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "email",
          "password",
          "company_name",
          "contact_name",
          "phone",
          "country"
        ],
        "title": "BuyerRegistrationRequest",
        "type": "object"
      },
      "ChangePasswordRequest": {
        "properties": {
          "current_password": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Current Password",
            "type": "string"
          },
          "new_password": {
            "maxLength": 128,
            "minLength": 12,
            "title": "New Password",
            "type": "string"
          }
        },
        "required": [
          "current_password",
          "new_password"
        ],
        "title": "ChangePasswordRequest",
        "type": "object"
      },
      "CurrentUserResponse": {
        "properties": {
          "company_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Name"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_login_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          },
          "must_change_password": {
            "title": "Must Change Password",
            "type": "boolean"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/AccountStatus"
          },
          "user_type": {
            "$ref": "#/components/schemas/UserType"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "user_type",
          "status",
          "must_change_password",
          "created_at"
        ],
        "title": "CurrentUserResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "identifier": {
            "maxLength": 320,
            "minLength": 1,
            "title": "Identifier",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "identifier",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "LogoutRequest": {
        "properties": {
          "refresh_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refresh Token"
          }
        },
        "title": "LogoutRequest",
        "type": "object"
      },
      "MessageResponse": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "MessageResponse",
        "type": "object"
      },
...
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refresh Token"
          }
        },
        "title": "RefreshRequest",
        "type": "object"
      },
      "RegistrationResponse": {
        "properties": {
          "company_name": {
            "title": "Company Name",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "message": {
            "default": "Registration received. Your account is pending administrator approval.",
            "title": "Message",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AccountStatus"
          },
          "user_type": {
            "$ref": "#/components/schemas/UserType"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "user_type",
          "status",
          "company_name"
        ],
        "title": "RegistrationResponse",
        "type": "object"
      },
...
      "TokenResponse": {
        "description": "Browser-safe auth response for login and refresh.\n\nCarries the bearer access token plus the complete current-user object the UI\nneeds to render immediately. The refresh token is deliberately NOT included:\nit is delivered only via the ``HttpOnly`` refresh cookie so it is never\nreadable by browser JavaScript.",
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "must_change_password": {
            "default": false,
            "title": "Must Change Password",
            "type": "boolean"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/CurrentUserResponse"
          }
        },
        "required": [
          "access_token",
          "expires_at",
          "user"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
...
      "UserType": {
        "enum": [
          "ADMIN",
          "BUYER",
          "VENDOR"
        ],
        "title": "UserType",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VendorRegistrationRequest": {
        "additionalProperties": false,
        "description": "Vendor sign-up: the buyer fields plus a non-empty supply-category list.",
        "properties": {
          "company_name": {
            "maxLength": 255,
            "minLength": 2,
            "title": "Company Name",
            "type": "string"
          },
          "contact_name": {
            "maxLength": 255,
            "minLength": 2,
            "title": "Contact Name",
            "type": "string"
          },
          "country": {
            "maxLength": 100,
            "minLength": 2,
            "title": "Country",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 12,
            "title": "Password",
            "type": "string"
          },
          "phone": {
            "maxLength": 50,
            "minLength": 3,
            "title": "Phone",
            "type": "string"
          },
          "supply_categories": {
            "items": {
              "maxLength": 255,
              "minLength": 2,
              "type": "string"
            },
            "maxItems": 50,
            "minItems": 1,
            "title": "Supply Categories",
            "type": "array"
          },
          "username": {
            "maxLength": 150,
            "minLength": 3,
            "pattern": "^[A-Za-z0-9._-]+$",
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "email",
          "password",
          "company_name",
          "contact_name",
          "phone",
          "country",
          "supply_categories"
        ],
        "title": "VendorRegistrationRequest",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Leo Infinity Trade Portal API \u2014 Phase 0 & Phase 1: foundation, authentication, account approval, lockout, refresh-token rotation, and RBAC.",
    "title": "Leo Infinity Trade Portal API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/change-password": {
      "post": {
        "description": "Change the caller's password.\n\nThis is **session-invalidating**: every active refresh session is revoked and\n``auth_version`` is bumped so all outstanding access tokens stop working. The\nrefresh cookie is cleared and the caller must authenticate again.",
        "operationId": "change_password_api_v1_auth_change_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Password changed. ALL active sessions are revoked (auth_version bumped) and the refresh cookie is cleared \u2014 the caller MUST log in again. The response body is `{\"message\": \"Password changed. Please log in again.\"}`."
          },
          "401": {
            "description": "Missing/invalid access token or wrong current password."
          },
          "409": {
            "description": "New password must differ from the current password."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Change own password (invalidates all sessions; requires re-login)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logout_api_v1_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/LogoutRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Logout",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RefreshRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/register/buyer": {
      "post": {
        "operationId": "register_buyer_api_v1_auth_register_buyer_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuyerRegistrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Buyer",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/register/vendor": {
      "post": {
        "operationId": "register_vendor_api_v1_auth_register_vendor_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VendorRegistrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Vendor",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/sessions/{session_id}": {
      "delete": {
        "operationId": "revoke_session_api_v1_auth_sessions__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Session",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}